Commit 2 Philosophers.
added the main philosopher loop. The data parsing is going well. Philosophers still get stuck in deadlock even though I have a mutex protecting it in the main loop, so work on that. I still have to implement the time_to_die aspect as i'm doing nothing with it as of now. Oh yeah and input validation and protection ofcourse. It is a decent start.